+++ openapi_builder/converters/parameter/base.py
@@ -1,25 +1,20 @@
 import typing
+from abc import ABC
 
 from openapi_builder.specification import Schema
 
-PARAMETER_CONVERTERS: typing.List["ParameterConverter"] = []
+if typing.TYPE_CHECKING:
+    from .manager import ParameterManager
 
 
-def register_parameter_converter(converter_class):
-    """Decorator for registering a converter.
-
-    The parameter converter is initialized without any arguments.
-    """
-    instance = converter_class()
-    PARAMETER_CONVERTERS.append(instance)
-    return converter_class
-
-
-class ParameterConverter:
+class ParameterConverter(ABC):
     """Converter for a certain class that returns an openapi_builder.specification.Schema."""
 
     converts_class = None
     """Specification of the class that it converts."""
+
+    def __init__(self, manager: "ParameterManager"):
+        self.manager: ParameterManager = manager
 
     def matches(self, value) -> bool:
         """Returns True if the Converter can match the specified class."""
@@ -27,4 +22,4 @@ class ParameterConverter:
 
     @property
     def schema(self) -> Schema:
-        raise NotImplementedError
+        raise NotImplementedError()